simply/simply/apps/app1/backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
pybind11==2.11.1
